refactor: :fire: move services crate into server for complete access to the type system
