+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["apps/server", "lib/tests", "lib/services"]
+members = ["apps/server", "tests"]
