+++ apps/server/Cargo.toml
@@ -14,13 +14,16 @@ required-features = ["seeder"]
 
 [dependencies]
 tokio = { version = "1.46.1", features = ["full"] }
+
 axum = "0.8.4"
 sword = { git = "https://github.com/MrRevillod/sword", branch = "dev" }
 
-services = { path = "../../lib/services" }
-
 chrono = { version = "0.4.41", features = ["serde"] }
+chrono-tz = "0.10.4"
+
 uuid = { version = "1.17.0", features = ["serde", "v4"] }
+tera = "1.20.0"
+lettre = "0.11.17"
 
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
@@ -34,7 +37,6 @@ sqlx = { version = "0.8.5", features = [
     "derive",
     "migrate",
 ] }
-
 sea-query-sqlx = { version = "0.8.0-rc.8", features = [
     "sqlx-postgres",
     "with-chrono",
@@ -57,7 +59,7 @@ bcrypt = "0.17.0"
 
 fake = { version = "4.3.0", optional = true }
 rand = { version = "0.9.1", optional = true }
-thiserror = "2.0.12"
+thiserror = "2.0.14"
 
 [dev-dependencies]
 axum-test = "17.3.0"
