+++ apps/server/src/lib.rs
@@ -97,6 +97,33 @@ pub mod shared {
             Ok(())
         }
     }
+
+    pub mod services {
+        pub mod errors;
+        pub mod hasher;
+        pub mod mailer;
+        pub mod printer;
+        pub mod templates {
+            mod context;
+            mod files;
+
+            pub use context::*;
+            pub use files::*;
+        }
+
+        pub mod event_queue {
+            mod publisher;
+            mod sender;
+            mod subscriber;
+
+            mod events;
+
+            pub use events::*;
+            pub use publisher::*;
+            pub use sender::*;
+            pub use subscriber::*;
+        }
+    }
 }
 
 pub mod container;
