+++ apps/server/src/practices/service.rs
@@ -1,6 +1,6 @@
+use crate::shared::services::event_queue::{Event, EventQueue};
 use async_trait::async_trait;
 use serde_json::json;
-use services::event_queue::{Event, EventQueue};
 
 use shaku::{Component, Interface};
 use std::sync::Arc;
@@ -107,8 +107,6 @@ impl PracticeService for PracticeServiceImpl {
         let (enrollment, student, practice) =
             self.enrollments.get_by_id(enrollment_id).await?;
 
-        dbg!(&practice);
-
         let mut practice =
             practice.ok_or(AppError::ResourceNotFound(*practice_id))?;
 
@@ -183,6 +181,6 @@ impl PracticeService for PracticeServiceImpl {
             .await?
             .ok_or(AppError::ResourceNotFound(*id))?;
 
-        self.practices.delete(id).await
+        self.practices.delete(&practice.id).await
     }
 }
