+++ apps/server/src/shared/services/errors.rs
