+++ apps/server/src/shared/services/event_queue/events.rs
