+++ apps/server/src/shared/services/mailer/templates/practice/approval/student.html
