+++ apps/server/src/shared/services/mailer/templates/practice/approval/teacher.html
