+++ apps/server/src/shared/services/mailer/templates/practice/creation/student.html
