+++ apps/server/src/shared/services/mailer/templates/practice/creation/supervisor.html
