+++ apps/server/src/shared/services/mailer/templates/practice/rejection/student.html
