+++ apps/server/src/shared/services/mailer/templates/practice/rejection/teacher.html
