+++ apps/server/src/shared/services/mailer/templates/system/welcome.html
