+++ apps/server/src/shared/services/printer/templates/authorization.typ
