+++ apps/server/src/shared/services/printer/templates/logo.png
