--- lib/services/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "services"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-tokio = { version = "1.46.1", features = ["full"] }
-sword = { git = "https://github.com/MrRevillod/sword", branch = "dev" }
-
-serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.142"
-tera = "1.20.0"
-
-lettre = "0.11.17"
-tracing = "0.1.41"
-
-shaku = { version = "0.6.2", features = ["thread_safe"] }
-tower-http = { version = "0.6.6", features = ["trace", "cors"] }
-tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-
-async-trait = "0.1.88"
-thiserror = "2.0.12"
-bcrypt = "0.17.0"
-chrono = "0.4.41"
-chrono-tz = "0.10.4"
