--- lib/services/src/lib.rs
@@ -1,24 +1,0 @@
-pub mod errors;
-pub mod hasher;
-pub mod mailer;
-pub mod printer;
-pub mod templates {
-    mod context;
-    mod files;
-
-    pub use context::*;
-    pub use files::*;
-}
-
-pub mod event_queue {
-    mod publisher;
-    mod sender;
-    mod subscriber;
-
-    mod events;
-
-    pub use events::*;
-    pub use publisher::*;
-    pub use sender::*;
-    pub use subscriber::*;
-}
